Move security config to one package
