+++ src/main/java/com/example/classroom/config/security/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.example.classroom.config;
+package com.example.classroom.config.security;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
